requirements.txt
noise
flask
Pillow